add tests and fix warnings

+++ tests/test_get_spin_isospin_wave_function.py
@@ -1,6 +1,7 @@
 from nuclear_qmc.spin.get_spin_isospin_wave_function import get_spin_isospin_wave_function
+import jax.numpy as jnp
 
 
 class TestGetWFC:
     def test_get_spin_isospin_wave_function(self):
-        computed = get_spin_isospin_wave_function(1, 1, True)
+        computed = get_spin_isospin_wave_function(1, 1, True, dtype=jnp.float32)
